requirements.txt
#Write dependecies below. They are case sensitive
python-dotenv
python-telegram-bot
six
openai